Blackjack Project
If the total of the cards in our hand is more than 21, it becomes a Bust and means we lose immediately. It is 6, 6 or 9, 9, but Vale, Kız and Papaz count as 10 each.
Ase either counts as 1 or counts as 11.
And if the b_cards total is <17, another trick must be taken.
In this game, b_cards = my cards, while k_cards = the other party's cards.